--- src/analysis/classic/paladin/CombatLogParser.ts
@@ -1,40 +1,0 @@
-import ManaTracker from 'parser/core/healingEfficiency/ManaTracker';
-import AbilityTracker from 'parser/shared/modules/AbilityTracker';
-import ManaLevelChart from 'parser/shared/modules/resources/mana/ManaLevelChart';
-import ManaUsageChart from 'parser/shared/modules/resources/mana/ManaUsageChart';
-import SpellManaCost from 'parser/shared/modules/SpellManaCost';
-import BaseCombatLogParser from 'parser/classic/CombatLogParser';
-import PreparationRuleAnalyzer from 'parser/classic/modules/features/Checklist/PreparationRuleAnalyzer';
-import lowRankSpellsSuggestion from 'parser/classic/suggestions/lowRankSpells';
-
-import lowRankSpells from './lowRankSpells';
-import Abilities from './modules/Abilities';
-import Buffs from './modules/Buffs';
-import Checklist from './modules/checklist/Module';
-import AlwaysBeCasting from './modules/features/AlwaysBeCasting';
-import HealingEfficiencyDetails from './modules/features/HealingEfficiencyDetails';
-import HealingEfficiencyTracker from './modules/features/HealingEfficiencyTracker';
-
-class CombatLogParser extends BaseCombatLogParser {
-  static specModules = {
-    abilities: Abilities,
-    buffs: Buffs,
-    spellManaCost: SpellManaCost,
-    abilityTracker: AbilityTracker,
-    manaLevelChart: ManaLevelChart,
-    manaUsageChart: ManaUsageChart,
-    alwaysBeCasting: AlwaysBeCasting,
-    preparationRuleAnalyzer: PreparationRuleAnalyzer,
-
-    // Mana Tab
-    manaTracker: ManaTracker,
-    hpmTracker: HealingEfficiencyTracker,
-    hpmDetails: HealingEfficiencyDetails,
-
-    checklist: Checklist,
-
-    lowRankSpells: lowRankSpellsSuggestion(lowRankSpells),
-  };
-}
-
-export default CombatLogParser;
