+++ src/analysis/classic/paladin/holy/CHANGELOG.tsx
@@ -1,0 +1,7 @@
+import { change, date } from 'common/changelog';
+import { jazminite } from 'CONTRIBUTORS';
+
+export default [
+  change(date(2023, 2, 12), 'Add Classic Holy Paladin stub.', jazminite),
+  change(date(2022, 11, 4), 'Update and re-org Classic Paladin spells.', jazminite),
+];
