+++ src/analysis/classic/paladin/holy/CombatLogParser.ts
@@ -1,0 +1,46 @@
+// Base file
+import BaseCombatLogParser from 'parser/classic/CombatLogParser';
+// Core
+import Haste from './modules/core/Haste';
+// Shared
+import lowRankSpellsSuggestion from 'parser/classic/suggestions/lowRankSpells';
+import { lowRankSpells } from '../shared';
+import ManaLevelChart from 'parser/shared/modules/resources/mana/ManaLevelChart';
+import ManaTracker from 'parser/core/healingEfficiency/ManaTracker';
+import ManaUsageChart from 'parser/shared/modules/resources/mana/ManaUsageChart';
+import SpellManaCost from 'parser/shared/modules/SpellManaCost';
+// Features
+import Abilities from './modules/features/Abilities';
+import AlwaysBeCasting from './modules/features/AlwaysBeCasting';
+import Buffs from './modules/features/Buffs';
+import CooldownThroughputTracker from './modules/features/CooldownThroughputTracker';
+import Checklist from './modules/checklist/Module';
+// Healer Features
+import HealingEfficiencyDetails from './modules/features/HealingEfficiencyDetails';
+import HealingEfficiencyTracker from './modules/features/HealingEfficiencyTracker';
+// Spells
+
+class CombatLogParser extends BaseCombatLogParser {
+  static specModules = {
+    // Core
+    haste: Haste,
+    // Shared
+    lowRankSpells: lowRankSpellsSuggestion(lowRankSpells),
+    manaLevelChart: ManaLevelChart,
+    manaTracker: ManaTracker,
+    manaUsageChart: ManaUsageChart,
+    spellManaCost: SpellManaCost,
+    // Features
+    abilities: Abilities,
+    alwaysBeCasting: AlwaysBeCasting,
+    buffs: Buffs,
+    cooldownThroughputTracker: CooldownThroughputTracker,
+    checklist: Checklist,
+    // Healer Features
+    hpmTracker: HealingEfficiencyTracker,
+    hpmDetails: HealingEfficiencyDetails,
+    // Spells
+  };
+}
+
+export default CombatLogParser;
