+++ src/analysis/classic/paladin/holy/index.ts
@@ -1,0 +1,2 @@
+// Reference this file in src\parser\index.ts
+export { default } from './CONFIG';
