+++ src/analysis/classic/paladin/holy/modules/checklist/Module.tsx
@@ -1,27 +1,40 @@
+// Core
+import BaseChecklist from 'parser/shared/modules/features/Checklist/Module';
+import Component from './Component';
+// Shared
 import CastEfficiency from 'parser/shared/modules/CastEfficiency';
 import Combatants from 'parser/shared/modules/Combatants';
-import BaseChecklist from 'parser/shared/modules/features/Checklist/Module';
+import CombatPotionChecker from 'parser/classic/modules/items/CombatPotionChecker';
 import ManaValues from 'parser/shared/modules/ManaValues';
 import PreparationRuleAnalyzer from 'parser/classic/modules/features/Checklist/PreparationRuleAnalyzer';
-
+// Features
 import AlwaysBeCasting from '../features/AlwaysBeCasting';
-import Component from './Component';
+// Spells
+// import SpellName from './modules/spells';
 
 class Checklist extends BaseChecklist {
   static dependencies = {
     ...BaseChecklist.dependencies,
-    combatants: Combatants,
+    // Shared
     castEfficiency: CastEfficiency,
+    combatants: Combatants,
+    combatPotionChecker: CombatPotionChecker,
     manaValues: ManaValues,
-    alwaysBeCasting: AlwaysBeCasting,
     preparationRuleAnalyzer: PreparationRuleAnalyzer,
+    // Features
+    alwaysBeCasting: AlwaysBeCasting,
+    // Spells
   };
 
-  protected combatants!: Combatants;
+  // Shared
   protected castEfficiency!: CastEfficiency;
-  protected preparationRuleAnalyzer!: PreparationRuleAnalyzer;
+  protected combatants!: Combatants;
+  protected combatPotionChecker!: CombatPotionChecker;
   protected manaValues!: ManaValues;
+  protected preparationRuleAnalyzer!: PreparationRuleAnalyzer;
+  // Features
   protected alwaysBeCasting!: AlwaysBeCasting;
+  // Spells
 
   render() {
     return (
@@ -30,11 +43,12 @@ class Checklist extends BaseChecklist {
         castEfficiency={this.castEfficiency}
         thresholds={{
           ...this.preparationRuleAnalyzer.thresholds,
-
           manaLeft: this.manaValues.suggestionThresholds,
-          nonHealingTimeSuggestionThresholds: this.alwaysBeCasting
-            .nonHealingTimeSuggestionThresholds,
+          nonHealingTimeSuggestionThresholds:
+            this.alwaysBeCasting.nonHealingTimeSuggestionThresholds,
           downtimeSuggestionThresholds: this.alwaysBeCasting.downtimeSuggestionThresholds,
+          // Spells
+          // spellName: this.spellName.uptimeSuggestionThresholds,
         }}
       />
     );
