+++ src/analysis/classic/paladin/holy/modules/features/AlwaysBeCasting.tsx
@@ -1,50 +1,29 @@
 import { t } from '@lingui/macro';
 import { formatPercentage } from 'common/format';
-import { Options } from 'parser/core/Analyzer';
 import { When } from 'parser/core/ParseResults';
 import CoreAlwaysBeCastingHealing from 'parser/shared/modules/AlwaysBeCastingHealing';
-
-import lowRankSpells, { LowRankSpells } from '../../lowRankSpells';
 import SPELLS from 'common/SPELLS/classic/paladin';
 
 class AlwaysBeCasting extends CoreAlwaysBeCastingHealing {
   static HEALING_ABILITIES_ON_GCD: number[] = [
-    SPELLS.HOLY_SHOCK.id,
+    // List of healing spells on GCD
     SPELLS.FLASH_OF_LIGHT.id,
+    ...SPELLS.FLASH_OF_LIGHT.lowRanks,
     SPELLS.HOLY_LIGHT.id,
-    SPELLS.JUDGEMENT_OF_LIGHT.id,
-    SPELLS.LAY_ON_HANDS.id,
-    SPELLS.CLEANSE.id,
-    SPELLS.PURIFY.id,
+    ...SPELLS.HOLY_LIGHT.lowRanks,
   ];
-
-  constructor(options: Options) {
-    super(options);
-
-    const maxRankSpells: number[] = AlwaysBeCasting.HEALING_ABILITIES_ON_GCD;
-
-    const lrs: LowRankSpells = lowRankSpells;
-    for (const spell_id of maxRankSpells) {
-      AlwaysBeCasting.HEALING_ABILITIES_ON_GCD = AlwaysBeCasting.HEALING_ABILITIES_ON_GCD.concat(
-        lrs[spell_id],
-      );
-    }
-    AlwaysBeCasting.HEALING_ABILITIES_ON_GCD = AlwaysBeCasting.HEALING_ABILITIES_ON_GCD.filter(
-      (id: number) => Boolean(id),
-    );
-  }
 
   suggestions(when: When) {
     const deadTimePercentage = this.totalTimeWasted / this.owner.fightDuration;
 
     when(deadTimePercentage)
-      .isGreaterThan(0.15)
+      .isGreaterThan(0.25)
       .addSuggestion((suggest, actual, recommended) =>
         suggest('Your downtime can be improved. Try to Always Be Casting (ABC).')
           .icon('spell_mage_altertime')
           .actual(
             t({
-              id: 'priest.holy.suggestions.alwaysBeCasting.downtime',
+              id: 'shared.suggestions.alwaysBeCasting.downtime',
               message: `${formatPercentage(actual)}% downtime`,
             }),
           )
