--- src/analysis/classic/paladin/index.ts
@@ -1,1 +1,0 @@
-export { default } from './CONFIG';
