+++ src/analysis/classic/paladin/shared/index.ts
@@ -1,0 +1,1 @@
+export { default as lowRankSpells } from './lowRankSpells';
