+++ src/analysis/classic/paladin/shared/lowRankSpells.ts
